Migrate git delete coauthor
